# Conflicts:
#	data/src/main/java/com/gdg/data/datasource/CrowdZeroDataSource.kt
#	domain/src/main/java/com/gdg/domain/repository/CrowdZeroRepository.kt

+++ data/src/main/java/com/gdg/data/datasource/CrowdZeroDataSource.kt
@@ -6,7 +6,7 @@ import com.gdg.data.dto.response.CongestionResponseDto
 import com.gdg.data.dto.response.WeatherResponseDto
 
 interface CrowdZeroDataSource {
-    suspend fun getWeather(areaId: Long): BaseResponse<WeatherResponseDto>
-    suspend fun getCongestion(areaId: Long): BaseResponse<CongestionResponseDto>
+    suspend fun getWeather(areaId: Int): BaseResponse<WeatherResponseDto>
+    suspend fun getCongestion(areaId: Int): BaseResponse<CongestionResponseDto>
     suspend fun getAssembly(date: String): BaseResponse<AssemblyResponseDto>
 }
