+++ domain/src/main/java/com/gdg/domain/repository/CrowdZeroRepository.kt
@@ -5,7 +5,7 @@ import com.gdg.domain.entity.ScheduleEntity
 import com.gdg.domain.entity.WeatherEntity
 
 interface CrowdZeroRepository {
-    suspend fun getWeather(areaId: Long): Result<WeatherEntity>
-    suspend fun getCongestion(areaId: Long): Result<CongestionEntity>
+    suspend fun getWeather(areaId: Int): Result<WeatherEntity>
+    suspend fun getCongestion(areaId: Int): Result<CongestionEntity>
     suspend fun getAssembly(date: String): Result<ScheduleEntity>
 }
